requirements.txt
requests>=2.31.0
flask>=3.0.0
urllib3>=2.0.0
schedule>=1.2.0
httpx>=0.28.0
playwright>=1.57.0
python-dotenv>=1.0.0
pymysql>=1.1.0

